requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
coverage==7.6.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==33.0.0
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
loguru==0.7.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomli==2.1.0
typing_extensions==4.12.2
uvicorn==0.24.0
